Move safepoint operations from debug-adapter.c to safepoints.h to share with command-line debugger
