+++ core/debug/debug-adapter.c
@@ -1,12 +1,10 @@
 #include <assert.h>
 #include <ctype.h>
 #include <errno.h>
-#include <inttypes.h>
 #include <pthread.h>
 #include <stdarg.h>
 #include <stdbool.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #ifdef PLATFORM_WINDOWS
   #include <fcntl.h>
@@ -18,6 +16,8 @@
   #include <unistd.h>
   typedef int SOCKET;
 #endif
+
+#include "safepoints.h"
 
 static inline void* memclear(void* data, size_t size) {
   return memset(data, 0, size);
@@ -117,69 +117,9 @@ static inline void free_path(const char* s) {
   free((char*)s);
 }
 
-enum {
-  NOP = 0x90,
-  INT3 = 0xCC
-};
-
-typedef const struct {
-  uint8_t* const address;
-  const uint64_t group;
-} SafepointAddress;
-
-typedef const struct {
-  const uint64_t length;
-  SafepointAddress addresses[];
-} AddressList;
-static void AddressList_write(AddressList* list, const uint8_t inst) {
-  for (SafepointAddress *p = list->addresses, *lim = p + list->length; p < lim; p++)
-    *p->address = inst;
-}
-static inline SafepointAddress* AddressList_find(AddressList* list, const void* pc) {
-  for (SafepointAddress *p = list->addresses, *lim = p + list->length; p < lim; p++)
-    if (p->address == pc)
-      return p;
-  return NULL;
-}
-
-typedef const struct {
-  const uint64_t line;
-  AddressList* const address_list;
-} SafepointEntry;
-static inline void SafepointEntry_write(SafepointEntry* entry, const uint8_t inst) {
-  AddressList_write(entry->address_list, inst);
-}
-static SafepointAddress* SafepointEntry_find(SafepointEntry* entry, const void* pc) {
-  return AddressList_find(entry->address_list, pc);
-}
-
-typedef const struct {
-  const uint64_t num_entries;
-  const char* const filename;
-  SafepointEntry entries[];
-} FileSafepoints;
-static void FileSafepoints_write(FileSafepoints* file, const uint8_t inst) {
-  for (SafepointEntry *entry = file->entries, *lim = entry + file->num_entries; entry < lim; entry++)
-    SafepointEntry_write(entry, inst);
-}
-static inline SafepointEntry* FileSafepoints_find(FileSafepoints* file, uint64_t line) {
-  if (file)
-    for (SafepointEntry *entry = file->entries, *lim = entry + file->num_entries; entry < lim; entry++)
-      if (entry->line >= line)
-        return entry;
-  return NULL;
-}
-
-typedef const struct {
-  const uint64_t num_files;
-  FileSafepoints* const files[];
-} SafepointTable;
 SafepointTable* app_safepoint_table;
 static void Safepoints_write(const uint8_t inst) {
-  SafepointTable* safepoints = app_safepoint_table;
-  if (safepoints)
-    for (uint64_t i = 0, num_files = app_safepoint_table->num_files; i < num_files; i++)
-      FileSafepoints_write(safepoints->files[i], inst);
+  SafepointTable_write(app_safepoint_table, inst);
 }
 static FileSafepoints* Safepoints_find_file(const char* file_name) {
   const char* file_path = get_absolute_path(file_name);
